requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-adk
google-adk[a2a]
google-genai